Refactor trading logic and enhance UI features
- Updated `ChannelInfo` to use `long` for `Id` property.
- Added new input parameters for price tolerance and advanced features in `CompleteTelegramEA.mq5`.
- Refactored `CheckForNewSignals` for improved readability and error handling; introduced `ValidateStopLevels`.
- Enhanced `ExecuteSingleTrade` with price tolerance checks and adjustments for SL/TP levels.
- Added `ValidateFinalStopLevels` to ensure valid SL/TP before trade execution.
- Introduced `FileSystemWatcher` and cleanup timer in `Form1.cs` for signal file monitoring.
- Updated UI with new buttons for clearing old signals and checking the signal file.
- Modified `ProcessedSignal` to use `DateTime.Now` for timestamps.
- Improved `ValidateSignal` method in `SignalProcessingService.cs` for SL/TP logic validation.
- Enhanced `TelegramService` with debug message handling and error handling during polling.
- Overall code structure and readability improvements through refactoring.

+++ TelegramEAManager/ChannelInfo.cs
@@ -8,7 +8,7 @@ namespace TelegramEAManager
 {
     public class ChannelInfo
     {
-        public int Id { get; set; }
+        public long Id { get; set; }  // Changed from int to long
         public string Title { get; set; } = "";
         public string Username { get; set; } = "";
         public string Type { get; set; } = "";
